Add option to show_urls to set the language
This is really useful if you use i18n_patterns, as if en-us is not in
settings.LANGUAGES you will get an error for show_urls along the lines of:

    Error occurred while trying to load urls: The active language en-us is
    not in the list of configured languages: [u'en', u'es']

So the option added is -l --language= which allows you to set the
language code.  It's useful to be able to check all sometimes, as with
some systems (eg django-cms) the URL itself is i18n'ed.