Resolve RemovedInDjango110Warning for OptParser and NoArgsCommand.
The BaseCommand OptParser and NoArgsCommand RemovedInDjango110Warnings
are really two sides of the same coin. We need to provide a
compatibility layer between OptParser and ArgParse, and NoArgsCommand
is a wrapper around BaseCommand that simply ensured no arguments were
added to the command. To update NoArgsCommand, by sub-classing
BaseCommand and changing handle_noargs() to handle(), in a Command that
does not add options, would simply expose the other deprecation.

To resolve these, we use a sub-class of BaseCommand to dynamically
generate an options list from the provided add_arguments() of a
sub-class. This way the API is the same when Django < 1.8 support is
removed and all that needs to change in the command files is an import.
Implementation wise, for BaseCommand, AppCommand, and LabelCommand the
compatibility classes are a drop-in while changing to the new format of
defining add_arguments() rather than appending to options. For
NoArgsCommand, we need to change handle_noargs() to handle() with the
appropriate function signature. When changing to parser.add_argument(),
we must also be sure that any 'type=' kwarg from make_option() is
converted to an actual type (e.g. str, int) rather than a string of the
type name.