sort the listdir of jobs, so that jobs will run in a predictable order
This makes it so that jobs will be run in alphabetic order per directory, starting with the jobs dir, then in order of time period from smallest (minutely) to largest (yearly), still contingent on the specification of time period as before. This allows the user to set the order of their jobs in a way similar to init; e.g. name the files with `00_<jobname>` for it to go before `10_<jobname>`